update: organize code in components folder: TODO move editor to creation form
