+++ marketplace/ui/src/components/MarketPlace/ProductDetail.js
@@ -41,7 +41,7 @@ import "react-responsive-carousel/lib/styles/carousel.min.css"; // requires a lo
 import { Carousel } from "react-responsive-carousel"
 import { Images } from "../../images";
 import ProductItemDetails from "./ProductItemDetails";
-import RichEditor from "./RichEditor";
+import RichEditor from "../RichEditor";
 
 const ProductDetails = ({ user, users }) => {
   const { state, pathname } = useLocation();
@@ -528,11 +528,12 @@ const ProductDetails = ({ user, users }) => {
                         <Paragraph
                           className="text-[#202020] text-sm"
                         >
-                          <RichEditor
-                            details={details}
-                            user={user}
-                            isAuthenticated={isAuthenticated}
-                          />
+                          {details?.description?.split('\n').map((line, index) => (
+                            <React.Fragment key={index}>
+                              {line}
+                              <br />
+                            </React.Fragment>
+                          ))}
                         </Paragraph>
                       </div>
                     ),
