+++ marketplace/ui/src/components/MarketPlace/index.css
@@ -86,118 +86,3 @@ img.ant-image-preview-img {
   width: 90%;
 }
 
-.tiptap {
-  > * + * {
-    margin-top: 0.75em;
-  }
-
-  ul, ol {
-    padding: 0 1rem;
-  }
-
-  h1, h2, h3, h4, h5, h6 {
-    line-height: 1.1;
-  }
-
-  code {
-    background-color: rgba(97, 97, 97, 0.1);
-    color: #616161;
-  }
-
-  pre {
-    background: #0D0D0D;
-    color: #FFF;
-    font-family: 'JetBrainsMono', monospace;
-    padding: 0.75rem 1rem;
-    border-radius: 0.5rem;
-
-    code {
-      color: inherit;
-      padding: 0;
-      background: none;
-      font-size: 0.8rem;
-    }
-  }
-
-  img {
-    max-width: 100%;
-    height: auto;
-  }
-
-  blockquote {
-    padding-left: 1rem;
-    border-left: 2px solid rgba(13, 13, 13, 0.1);
-  }
-
-  hr {
-    border: none;
-    border-top: 2px solid rgba(13, 13, 13, 0.1);
-    margin: 2rem 0;
-  }
-
-  button {
-    padding: 0.5rem 1rem;
-    margin: 0.25rem;
-    background-color: white;
-    color: #616161;
-    border: 1px solid rgba(97, 97, 97, 0.5);
-    border-radius: 4px;
-    cursor: pointer;
-    font-family: 'Arial', sans-serif;
-    font-size: 0.875rem;
-
-    &:hover {
-      background-color: rgba(97, 97, 97, 0.1);
-    }
-
-    &.is-active {
-      background-color: #0D0D0D;
-      color: white;
-    }
-  }
-}
-
-.editor-container {
-  padding: 20px; /* Adjust padding as needed */
-  border: 1px solid #ccc; /* Optional: adds a border around the editor */
-  margin: auto; /* Center the editor in the page */
-  width: 80%; /* Adjust width as needed */
-  min-height: 400px; /* Adjust minimum height as needed */
-}
-
-/* Additional style to increase the inner padding of the contenteditable area */
-.ProseMirror {
-  padding: 20px; /* Inner padding */
-  min-height: 100%; /* Ensures the padding applies even when the editor is empty */
-}
-
-.toolbar-color-picker {
-  position: relative;
-  display: inline-block;
-  margin: 0;
-}
-
-
-
-.bubble-menu {
-  display: flex;
-  background-color: #0D0D0D;
-  padding: 0.2rem;
-  border-radius: 0.5rem;
-
-  button {
-    border: none;
-    background: none;
-    color: #FFF;
-    font-size: 0.85rem;
-    font-weight: 500;
-    padding: 0 0.2rem;
-    opacity: 0.6;
-
-    &:hover,
-    &.is-active {
-      opacity: 1;
-    }
-  }
-}
-
